HttpContextExtension which create a correct BadRequest json. Also ssed in PlantValidatorMiddleware
