+++ src/Equinor.Procosys.Preservation.WebApi/Middleware/PlantValidatorMiddleware.cs
@@ -1,6 +1,8 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using Equinor.Procosys.Preservation.Domain;
 using Equinor.Procosys.Preservation.MainApi.Plant;
+using Equinor.Procosys.Preservation.WebApi.Misc;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
@@ -25,11 +27,11 @@ namespace Equinor.Procosys.Preservation.WebApi.Middleware
             {
                 if (!await plantCache.IsAValidPlantAsync(plantId))
                 {
-                    var error = $"Plant '{plantId}' is not a valid plant";
-                    logger.LogError(error);
-                    context.Response.StatusCode = 400;
-                    context.Response.ContentType = "application/text";
-                    await context.Response.WriteAsync(error);
+                    var errors = new Dictionary<string, string[]>
+                    {
+                        {CurrentPlantMiddleware.PlantHeader, new[] {$"Plant '{plantId}' is not a valid plant"}}
+                    };
+                    await context.WriteBadRequestAsync(errors, logger);
                     return;
                 }
             }
